Update name of course and sections
This script can be used to update the name of the course and sections. The CANVAS_API_KEY and CANVAS_API_URL are passed as environment variables.

The current version requires a JSON be downloaded for courses to be updated. This allows for it to be repeated easily as well as have a backup list of the courses modified. It's easy to get 200 courses back from a search on a page. This would need modification to retrieve more.

Currently it's hardcoded for one specific case of replacing HBEHED with HBEHEQ.

This was generated with assistance of GPT.